refactor: move RegisterAgent to AgentRunner
